requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==4.0
django-debug-toolbar==3.2.2
django-markdown==0.8.4
django-taggit==2.0.0
importlib-metadata==4.8.2
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.1
zipp==3.6.0
